ES3-0.5.9 | ES5-0.5.7
ES5 0.5.7
  - `sleipnir.EventEmitter.emit` now uses an invoker. You can request
the "free" argument `$event` (holding the event type) as an handler
argument. This could reveal useful in the context of an eventHandler
object, which may be unaware of that information otherwise.

ES3 0.5.9
  - `sleipnir.EventEmitter.emit` now uses an invoker. You can request
the "free" argument `$event` (holding the event type) as an handler
argument. This could reveal useful in the context of an eventHandler
object, which may be unaware of that information otherwise.